requirements.txt
selenium>=4.0.0
requests>=2.28.0
openpyxl>=3.1.0
pyyaml>=6.0
sendgrid>=6.10.0
webdriver-manager>=3.8.5
